Packagificationize

+++ run.sh
@@ -1,10 +1,21 @@
 #!/usr/bin/env bash
 # Never print SECRETS or TOKENS.
 
-DOCKER_PATH_TARGET_DIR="/home/selery/runningrepo/"
-RESULT_DIR="/home/electron/"
+OPENSELERY_TARGET_PROJECT=$1
+if [ -z ${OPENSELERY_TARGET_PROJECT} ]; then 
+    echo "Please provide the directory path of the project you want to use with openselery!"
+    echo "run.sh <some_project_dir>"
+    exit
+fi
+
+
 DOT_DIR="~/.openselery/"
 DOT_DIR="${DOT_DIR/#\~/$HOME}"
+RESULT_DIR="${DOT_DIR/results}"
+DOCKER_PATH_TARGET_DIR="/home/selery/runningrepo"
+
+### opening permission for whoever wants to write here (docker container)
+chmod 777 $RESULT_DIR
 
 # Mount the argument folder into the container \
 docker run --rm -t \
@@ -12,11 +23,14 @@ docker run --rm -t \
 --env LIBRARIES_API_KEY=$LIBRARIES_API_KEY \
 --env COINBASE_TOKEN=$COINBASE_TOKEN \
 --env COINBASE_SECRET=$COINBASE_SECRET \
--v $@:$DOCKER_PATH_TARGET_DIR \
--v $(realpath $DOT_DIR/results):/home/selery/results \
+-v $OPENSELERY_TARGET_PROJECT:$DOCKER_PATH_TARGET_DIR \
+-v $(realpath $RESULT_DIR):/home/selery/results \
 -v $(realpath $DOT_DIR/config):/home/selery/config \
 -u $(id -u $USER):$(id -g $USER) \
-openselery \
---config $DOCKER_PATH_TARGET_DIR/selery.yml --directory $DOCKER_PATH_TARGET_DIR/ --result /home/selery/results --tooling /home/config/tooling_repos.yml
+  openselery \
+    --config $DOCKER_PATH_TARGET_DIR/selery.yml \
+    --directory $DOCKER_PATH_TARGET_DIR \
+    --result /home/selery/results \
+    --tooling /home/config/tooling_repos.yml
